+++ src/ASDF_f.c
@@ -79,7 +79,7 @@ void asdf_create_stations_group_f_(hid_t *loc_id, char *station_name,
 }
 
 void asdf_define_waveform_f_(hid_t *loc_id, int *nsamples,
-                          int *start_time, double *sampling_rate,
+                          long long int *start_time, double *sampling_rate,
                           char *event_name, char *waveform_name,
                           int *data_id) {
   *data_id = ASDF_define_waveform(*loc_id, *nsamples, *start_time,
@@ -88,7 +88,7 @@ void asdf_define_waveform_f_(hid_t *loc_id, int *nsamples,
 
 
 void asdf_define_waveforms_f_(hid_t *loc_id, int *num_waveforms, int *nsamples,
-                              int *start_time, double *sampling_rate,
+                              long long int *start_time, double *sampling_rate,
                               char *event_name, char **waveform_names,
                               int *data_id, int *err) {
   *err = ASDF_define_waveforms(*loc_id, *num_waveforms, *nsamples,
